+++ packages/create-bison-app/tasks/copyFiles.js
@@ -42,14 +42,8 @@ async function copyFiles({ variables, targetFolder }) {
     copyWithTemplate(fromPath("_.env.ejs"), toPath(".env"), variables),
 
     copyWithTemplate(
-      fromPath("_.env.local.ejs"),
-      toPath(".env.local"),
-      variables
-    ),
-
-    copyWithTemplate(
       fromPath("_.env.development.ejs"),
-      toPath(".env.local"),
+      toPath(".env.development"),
       variables
     ),
 
@@ -67,7 +61,13 @@ async function copyFiles({ variables, targetFolder }) {
 
     copyWithTemplate(
       fromPath("_.env.test.local.ejs"),
-      toPath(".env.test"),
+      toPath(".env.test.local"),
+      variables
+    ),
+
+    copyWithTemplate(
+      fromPath("docker-compose.yml.ejs"),
+      toPath("docker-compose.yml"),
       variables
     ),
 
@@ -114,6 +114,7 @@ async function copyFiles({ variables, targetFolder }) {
         "prettier.config.js",
         "tsconfig.json",
         "tsconfig.cjs.json",
+        "config.ts",
       ],
       targetFolder,
       {